remove commented code and small refactor
