--- src/entity/mod.rs
@@ -1,1 +1,0 @@
-pub mod multi_select;
